requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
google-genai>=1.0.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
httpx>=0.27.0
pandas>=2.0.0
numpy>=1.24.0
librosa>=0.10.0
scikit-learn>=1.3.0
soundfile>=0.12.0
spotifyscraper>=2.1.0
melon-chart.py>=1.0.0
bugs-chart.py>=1.0.0
billboard.py>=7.0.0
